# Add new API and speedup notice sending.