```
benchmark                 old ns/op     new ns/op     delta
BenchmarkSendNotice-4     32225         31622         -1.87%

benchmark                 old allocs     new allocs     delta
BenchmarkSendNotice-4     122            120            -1.64%

benchmark                 old bytes     new bytes     delta
BenchmarkSendNotice-4     8586          7739          -9.86%
```