@vmihailenco 2016-03-25
Added benchmark to the PR description, but it is too noisy. Also renamed NotifyPanic to NotifyOnPanic.

>Can you please point this out?

Point what? I added here `WaitAndClose` and `NotifyOnPanic`.